fix(config): rename deprecated ruff TCH rule code to TC; normalize trailing EOF newlines
fractalmusic/pyproject.toml's ruff select list used "TCH" (flake8-type-
checking), which ruff itself has since renamed to "TC" — current ruff no
longer resolves it (`ruff rule TCH001` errors, suggesting TC001), and it's
part of why validate-pyproject's schema rejected the whole [tool.ruff.lint]
table. Following this project's own greenfield-first rule (no legacy
aliases), renamed to the current code rather than keeping the deprecated
one working.

Also includes the trailing-EOF-newline fixes pre-commit-hooks made on its
first real run across docs/assets/*.svg, scripts/*.py, two SPEC docs,
web/public/favicon.svg, and several web/src/gatople components — none of
these had been touched by this session's edits before; they're pre-existing
hygiene gaps this was the first tool run to actually catch. Bundled with
.pre-commit-config.yaml itself since pre-commit refuses to commit anything
while its own config is unstaged.

+++ scripts/produce_flamenco_song.py
@@ -88,7 +88,7 @@ def note_to_midi(note: str) -> int:
 
 
 def note_to_freq(note: str) -> float:
-    return 440.0 * (2.0 ** ((note_to_midi(note) - 69) / 12.0))
+    return float(440.0 * (2.0 ** ((note_to_midi(note) - 69) / 12.0)))
 
 
 def add_stereo(buf: np.ndarray, mono: np.ndarray, start_s: float, gain: float, pan: float) -> None:
@@ -126,7 +126,7 @@ def plucked_note(
     t = np.arange(length, dtype=np.float32) / SR
     partials = np.zeros(length, dtype=np.float32)
     for harmonic in range(1, 9):
-        gain = (brightness ** harmonic) / harmonic
+        gain = (brightness**harmonic) / harmonic
         partial_decay = np.exp(-t * decay * (0.45 + harmonic * 0.12))
         phase = RNG.uniform(0, math.tau)
         partials += gain * np.sin(math.tau * freq * harmonic * t + phase) * partial_decay
@@ -539,7 +539,11 @@ def main() -> None:
     }
     meta_path.write_text(json.dumps(metadata, indent=2), encoding="utf-8")
 
-    print(json.dumps({**metadata, "absolute_wav": str(wav_path), "absolute_midi": str(midi_path)}, indent=2))
+    print(
+        json.dumps(
+            {**metadata, "absolute_wav": str(wav_path), "absolute_midi": str(midi_path)}, indent=2
+        )
+    )
 
 
 if __name__ == "__main__":
